fixed our error when I building login_page
