+++ tests/utils/asserts.py
@@ -1,25 +1,29 @@
 import pytest
+import test
+from tests.utils.locators import Locators
 from time import sleep as wait
 from selenium.webdriver.remote.webdriver import WebDriver
-dataType = str | int | float | bool | list | tuple | set | dict
-
+from typing import Union
+from tests.utils.locators import Locators
+from typing import Tuple
+dataType = Union[str, int, float, bool, list, tuple, set, dict]
+Test = Tuple[WebDriver, Locators] 
 
 class Expect:
     # * Definición de funciones para realizar Asserciones con valores dados
     def __init__(self, actualValue: dataType):
-        if not isinstance(actualValue, dataType):
+        if not isinstance(actualValue, (str, int, float, bool, list, Tuple, set, dict)):
             raise ValueError(f'DataType not supported. try: {dataType}')
         self.value = actualValue
-        self.dataType = dataType
 
     def toBeEqual(self, expectedValue: dataType):
-        if not isinstance(expectedValue, dataType):
+        if not isinstance(expectedValue, (str, int, float, bool, list, tuple, set, dict)):
             raise ValueError(f'DataType not supported. try: {dataType}')
         assert self.value == expectedValue
         wait(1)
 
     def toNotBeEqual(self, expectedValue: dataType):
-        if not isinstance(expectedValue, dataType):
+        if not isinstance(expectedValue, (str, int, float, bool, list, tuple, set, dict)):
             raise ValueError(f'DataType not supported. try: {dataType}')
         assert self.value != expectedValue
         wait(1)
@@ -33,11 +37,11 @@ class Expect:
     def isTrue(self):
         if not isinstance(self.value, bool):
             raise ValueError(f'Value type not supported. Use Boolean type')
-        assert self.value == True
+        assert self.value is True
         wait(1)
 
     def isFalse(self):
         if not isinstance(self.value, bool):
             raise ValueError(f'Value type not supported. Use Boolean type')
-        assert self.value == False
+        assert self.value is False
         wait(1)
